% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{latex_escape}
\alias{latex_escape}
\title{Échappe les underscores pour LaTeX}
\usage{
latex_escape(x)
}
\arguments{
\item{x}{Chaîne de caractères (ex. nom de paramètre "beta_Q_VB").}
}
\value{
La chaîne avec les "_" échappés en "\\_".
}
\description{
Échappe les underscores pour LaTeX
}
\examples{
latex_escape("beta_Q_VB")
}
