% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{param_quantiles}
\alias{param_quantiles}
\title{Quantiles et moyenne d'un paramètre CODA}
\usage{
param_quantiles(param, probs = seq(0, 1, 0.1))
}
\arguments{
\item{param}{Un objet \code{mcmc} issu de \code{\link[coda]{read.coda}}.}

\item{probs}{Vecteur de probabilités pour les quantiles. Défaut : \code{seq(0, 1, 0.1)}.}
}
\value{
Une liste avec \code{summary} (sortie de \code{summary.mcmc}), \code{quantiles} et \code{mean}.
}
\description{
Calcule les quantiles (par défaut 5/15/.../95\%) et la moyenne d'un paramètre
CODA, dans le format attendu par les graphiques boxplot par année
(\code{\link{custom_plot}} / \code{\link{boxplot_years_png}}).
}
