% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{standardize_residuals}
\alias{standardize_residuals}
\title{Standardise des résidus par un écart-type tiré par itération MCMC}
\usage{
standardize_residuals(res, sigma)
}
\arguments{
\item{res}{Objet coda/matrice des résidus (itérations x années).}

\item{sigma}{Vecteur (1 valeur par itération MCMC) de l'écart-type associé.}
}
\value{
Un objet \code{mcmc} des résidus standardisés, mêmes noms de colonnes que \code{res}.
}
\description{
Calcule \code{std[i, t] = res[i, t] / sigma[i]}. Motif récurrent pour les
résidus standardisés du modèle (res_vichy_std, res_p_*_std,
res_wild_moy_*_std, res_juv_moy_*_std, ...).
}
